Pulled out inner class into own file
